Include a submitter entity in seeded notices
Supporting this required changes to nested validations on entity notice
roles.

Since all nested records are validated before any are inserted, earlier
entities are not present when later entities are validated and any
duplicated entities are not found. To fix this, we explicitly save the
entities as part of validation.