Create LegacyCsv::ErrorHandler
In the case of import error, the offending CSV row and its original
files are copied to a second directory.

After the import completes, this second directory not only contains all
information relevant to the attempted import and specific failure, but
is also itself importable if/when the ingestor is updated.