Import Original and Supporting files separately
This change modifies the import side to assume that the export side
knows which files are originals and which are supporting and exports
them as two distinct columns.

This allows us to attach them as the right kind of uploads preserving
the public visibility (i.e. ReadLevel) of the files as they were in the
old system.

We attempt to parse works out of *all* files (because, why not?) and
stop upon first successful parse. Typically, we'll either find works in
the first original file or not find anything at all.