Include url_originals in the search index
When querying from the API the results from a search query only include
what is in the elasticsearch index as opposed to what is in the model.
This update includes the url_originals in the search index. A subsequent
commit will then require a researcher token to see url_originals.