Allow removal of nested entity notice roles
Rails admin was not allowing users to delete an entity notice role. This would cause a problem if an entity was associated in multiple roles for a notice and a user wanted to change the name/assocation for one entity but not the other one. This change allows the admin user to sever the relationship and then create a new entity to make this kind of change.