Fix broken duplicate url tests
For unknown reasons when running the tests with FactoryGirl the two modified tests fail due to a Postgres unique violation. The difference between Rails 3.2 and 4.2 is the appearance of the following at the end of the Factory creation:

  SQL (0.4ms)  INSERT INTO "copyrighted_urls" ("url", "url_original", "created_at", "updated_at") VALUES ($1, $2, $3, $4) RETURNING "id"  [["url", "http://www.example.com/copyrighted"], ["url_original", "http://www.example.com/copyrighted"], ["created_at", "2016-07-25 19:08:48.450078"], ["updated_at", "2016-07-25 19:08:48.450078"]]

Manual testing of the functionality in rails console revealed the functionality works as expected, but the test does not. Switching to pure ruby instead of abstracting in FactoryGirl lead to getting these tests to pass for now.