Reload connections before suite to prevent SocketErrors
If you run a large number of tests in succession, you may end up with
most of them failing with a SocketError. It seems that this is the
result of Elasticsearch exhausting its connections. This may be
the result of __build_connections getting passed a nil array, resulting
in nil hostname/servname getting ultimately passed to getaddrinfo(),
but I've been unable to trace the specifics of why. (Maybe
sniffer.hosts returning an empty array if Elasticsearch is exhausted?
Maybe something about thread safety, what with webrick and capybara
running in separate threads?)

At any rate, the reload fixes this problem.