Remove line which caused intermittent test failures
The current_driver setter was run only once, at the beginning of the
test. This caused tests to have different behavior depending on whether
they were the first example run.

Removing this line means that instead Capybara uses the default driver
(RackTest) except when tests have a js: true flag, requiring
Poltergeist. RackTest is faster, so it's good to use it where possible.