Don't reimplement sign_in function multiple times
Warden supplies a login_as, which is faster than specifying the
Capybara steps (albeit slightly less faithful to real-world conditions).
This commit wraps it in a local function (since we only ever need
user scope, there's no need to write that everywhere). It then uses that
local function everywhere rather than repeatedly defining it.

This requires a bit of surgery for the PageObject descendants, since
they include user objects, and the sign_in function properly applies to
those internal objects and not to the PageObject itself. The associated
tests have been slightly rewritten to clarify when Users or PageObject
subclasses are being referenced.