Upgrade paperclip to 4.X
This requires being explicit about the fact that we're not validating
filetypes (or else actually validating them).

There are other things we could change with the 3->4 transition, but
it's not worth dumping substantial time into this since when we get to
Rails 5.2 paperclip is deprecated in favor of activestorage anyway.