Add turnout back in
Having a maintenance mode would be really helpful during deploys, and
this is far and away the recommended gem for it. Maybe turnout
conflicts with rspec but is not otherwise a problem? If that's the case,
we can get rspec to run by avoiding loading turnout, and then check on
a server to see if everything still works as hoped.