Add NoticeSubmissionFinalizer
This allows us to add the works to a Notice separately from the
original notice creation process. This is important because it is the
slow part of the process, and therefore we should do it outside of the
request/response loop. In a subsequent PR we will add Sidekiq to allow
for the work addition to be done asynchronously.