Speed up RailsAdmin::MainController#edit
This view performs a SELECT COUNT (*) on Notice. I can't track down
the exact location in the code - it's not inside of RailsAdmin - but
from reading the issues queue it looks like that count is performed
as a result of the has_many :notices relations, so by excluding
Notice from Topic and Entity we may be able to prevent it from
happening.

Testing on localhost indicates that the exclude_fields version is on
average about 14% faster; hopefully this has much more impact on prod,
where there are about two thousand times as many notices.