Change spec type to force Elasticsearch to run
The ultimate goal is to only spin up ES during feature tests, so that
non-feature tests can run quickly. SearchesModels depends on ES, though.
In time we need to mock out this dependency and switch this back to a
model type (or move things to spec/integration where they really are
integration tests), but for now, let's make sure the tests can pass.