Speed up admin by removing nested forms
Rendering the nested has_many page is by far the slowest part of the
operation, even on localhost with small data sets (and it gets slower
as things scale). Admins will still be able to add/remove works from
the notice page - they need to go to the work page itself to edit the
work, but that's what we already do.