Refactor SearchesModels
The core of many of our performance problems is that
SearchesModels#search instantiates an ActiveRecord object for every
elasticsearch result. This can be up to 20,000 Notices, each of which
instantiates a variety of related objects, resulting in millions of
allocations per search.

SearchesModels#search should have a smaller scope of responsibility -
just searching - and it should be up to callers to decide how they
want to postprocess the data (and also to do so in a way which limits
the work to what they actually need).

In subsequent commits #search will be altered to achieve this. First,
I'm refactoring it both to make it cleaner and to understand what it's
doing.

This commit should result in no functional changes.