Postprocess search results in TopicsHelper
This lets us pull only the metadata we need to display notice links,
instead of instantiating large numbers of AR objects just to populate
a sidebar.

We need to stop using the notice partial since we no longer have a
notice object. This runs a risk that the sidebar styling will drift
out of sync with similar styling elsewhere on the site. I may track
that down in a future commit if it's not too much work.