Make Notices::SearchController more performant
Because SearchesModels no longer is responsible for postprocessing data,
Notices::SearchController needs to handle the elasticsearch results.
Rather than instantiating a new NoticeSearchResult (and possibly also a
new Notice) for each ES result, we pull the already-existing Notice from
the database and temporarily enrich it with metadata for display.

This improved responsibility model made the logic inside of
Notices::SearchController#index cleaner, which made it possible to pull
duplicated logic out of format.html and format.json and streamline the
whole function.

This commit also moves some tests of the associated functionality from
spec/views to spec/integration; because elasticsearch results are needed
to populate the search index it turns out to be much easier to write
those tests cleanly inside the integration suite rather than mocking
things out in views tests.

This reduces the total number of instantiations by about 75% on
localhost (will probably be better on prod).