Refactor ReindexRun
This will make it easier for us to apply automated work redaction
via the existing cron job. It also lets us work with ReindexRun
instances rather than classes inside of index_changed_model_instances,
which allows for slightly more idiomatic ruby. (The interface between
ReindexRun and the rest of the world is still via the class, to avoid
having to do too much work disentangling existing code and cron jobs.)

This also updates a couple of tests that didn't actually do anything
so that they do in fact test the code they claim to test.