Improve performance of work redaction rake task
There are 203M works in the database, which raises several issues:
1) we don't want to put all of them into a loop to redact - the memory
   usage would be a nightmare and the odds of a system crash partway
   through taking the process down would be nontrivial
2) we do want to make sure that rerunning the rake task doesn't
   reprocess already-processed works, but instead moves ahead to works
   we haven't examined;
3) we DEFINITELY do not want to perform 203M notice reindexes, so we
   should make sure that we only reindex notices when we've changed
   their associated works.

We accomplish #1 with the limit query. (Extrapolating from local
benchmarking, it should take a little over an hour to run this cron
job; the database query on prod takes visible time, but not much.)

We accomplish #2 with the order query. This ensures that we start with
the earliest notices that haven't been touched yet by this process.
Anything already redacted will have also been updated, so we won't
catch it in the query, but we will iterate through everything in order.

We accomplish #3 by tracking works that have actually been redacted,
and only touching notices connected to these works.