Make railsadmin more performant
The lines which set xhr and cacheAll in the original include
`field.associated_collection_cache_all`, which for some reason does a
SELECT COUNT(*), which is outrageously time-consuming on the Works
table. Instead, we'll set those values to reasonable defaults and skip
the whole nightmare-world of SELECT COUNT(*) on hundreds of millions of
objects.