Try to handle intermittent failures
The past commit's tests succeeded on localhost but failed on travis.
what. And then when run with the same seed, failed on localhost, but
with a different failure. WHAT. I can't see any reason endogenous to the
individual tests for that to happen, so I'm looking for linkages
between tests. A failure to wipe the elasticsearch index would mess
with search results even if the db is properly reset.