Fix failing spec
Because Notice.recipient gets the *first* recipient of the list,
it misses the recipient we created, which is the *second*; the
first is created in the factory. This was not the case with prior
versions of factorybot because validations didn't run, but they do
now, so we must create a recipient to avoid failing the spec.