Test CMS pages
This required being able to set up a CMS for testing purposes. It turns
out the recommended way to do that is by using the importer -- but
once we've got that, we may as well also use it to initialize the real
site!

This required some yakshaving:
* update minor version of Comfy (the exporter is broken in 2.0.18)
* export desired config from localhost to seeds files
* add rake task for importing from seeds
* break out blog post importer into separate class, so it can also be
  used by the test file to create sample content

This means also writing tests of the migration process, not just the
CMS content.

Surprisingly much work to get to the first working test, but already
paying off in terms of writing additional tests and making sure the
test CMS is like the real one. Will pay off further on deployment.