Add layout for Pages
There doesn't seem to be a way to pass locals to our parent view from
a comfy layout, other than by rendering everything into a partial.
We still need to define the content fragments in the layout so that
they will be solicited in the page editor, but we can't render them
in the layout as we will be rendering them in the partial.