Refactor notice creation
A lot of the difficulty we were having, e.g. with form-style vs JSON
parameters, was because we were getting away from a Rails-y way of
doing things -- editing parameters directly rather than instantiating
objects and using the available syntactic sugar. This also led to code
that was hard to read.

This refactors our earlier notice creation process to get rid of the
NoticeSubmissionInitializer and NoticeSubmissionFinalizer objects.
Instead, a NoticeBuilder builds a raw notice object and does the edits
we require (e.g. to set defaults). While building does potentially
require instantiating a lot of objects, it is still dramatically faster
than persisting to the database, even with a large number of URLs.
Further, it lets us take advantage of built-in ActiveRecord validations,
rather than having to manually inspect parameters.

This refactor gets us closer to a natural rails style of doing things.

Next up: we will add additional validations to the Url objects in order
to avoid certain types of data loss.