Add flexbox styles for nav
If the number of items in the nav bar can vary without a deploy
(because of CMS changes), we need a stylesheet that is agnostic as to
how many items are in the header/footer. This is a good excuse for
flexbox!

Grafting the dropdown styles into it is ugly (and rewriting all of
those to flex is out of scope), but it turns out flexbox is great.