Organize Notice and Entity
There's so much stuff in both of these models that it's hard to tell
what we've already done. (It turns out we had already implemented a
version of the thing I'm planning to implement in the next commit --
except that it hasn't been working! Good to know.)

Based on https://www.zmwolski.com/Organizing-Ruby-on-Rails-Models, but:
1) I removed scopes since we don't use a ton of them;
2) I needed to put relationships earlier to make the code run -- some of
   the other categories rely on things we define as relationships